Code for finding inverse of matrix
The code shown here is used for finding inverse of a matrix. The initial
portion allows users to input the values of the required matrix. The
latter half checks if the calculation for given matrix has been done
earlier, to save time. If not, it performs calculation and stores result
in a variable for future reference.